Refactor: 채팅 관련 컴포넌트 리팩토링
