+++ src/components/chat/modules/ChatRoomList.styled.ts
@@ -1,4 +1,3 @@
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 export const SearchBar = styled.div`
@@ -40,45 +39,6 @@ export const SearchInput = styled.input`
   }
 `;
 
-export const PreviewTitle = styled.div`
-  display: flex;
-  align-items: baseline;
-  gap: 0.4375rem;
-`;
-
-export const PreviewContent = styled.div`
-  overflow: hidden;
-`;
-
-export const ProfileName = styled.span`
-  font-size: 1rem;
-  font-weight: 700;
-`;
-
-export const PreviewTime = styled.span`
-  color: #3c3c3c;
-  font-size: 0.8125rem;
-`;
-
-export const PreviewMessage = styled.p`
-  font-size: 0.8125rem;
-  color: #3c3c3c;
-  margin-top: 8px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-`;
-
-export const ProfileImage = styled.img`
-  object-fit: cover;
-  border-radius: 50%;
-`;
-
-export const ProfileImageWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
 export const Container = styled.div<{ $isMobile: boolean }>`
   width: ${({ $isMobile }) => ($isMobile ? '100%' : '410px')};
   border-right: 1px solid #eaeaea;
@@ -86,16 +46,4 @@ export const Container = styled.div<{ $isMobile: boolean }>`
   box-sizing: border-box;
   display: flex;
   flex-direction: column;
-`;
-
-export const ChatRoomPreviewContainer = styled.li<{ $isCurrentRoom: boolean }>`
-  padding: 1.25rem;
-  background-color: ${({ $isCurrentRoom }) => ($isCurrentRoom ? '#ffe6e6' : 'white')};
-  border-bottom: 1px solid #eaeaea;
-  box-sizing: border-box;
-`;
-
-export const ChatRoomPreviewLink = styled(Link)`
-  display: flex;
-  gap: 11px;
 `;
