+++ src/components/chat/modules/ChatRoomList.tsx
@@ -1,33 +1,10 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { AiFillCloseCircle } from 'react-icons/ai';
 import { BiSearchAlt2 } from 'react-icons/bi';
-import { useSearchParams } from 'react-router-dom';
-import { leaveChatRoom } from '../../../api/chat';
-import defaultProfileImgUrl from '../../../assets/basic-user-profile-img.png';
 import useChatContext from '../../../hooks/useChatContext';
-import { RoomWithLastMessageType } from '../../../supabase/database.types';
 import * as S from './ChatRoomList.styled';
 import { useViewport } from '../../../hooks';
-
-const calculateTimeDifference = (inputTime: string): string => {
-  const inputDate = new Date(inputTime);
-  const currentTime = new Date();
-
-  const timeDifference = currentTime.getTime() - inputDate.getTime();
-  const timeDifferenceInMinutes = Math.floor(timeDifference / (1000 * 60));
-
-  if (timeDifferenceInMinutes < 1) {
-    return '방금 전';
-  } else if (timeDifferenceInMinutes < 60) {
-    return `${timeDifferenceInMinutes}분 전`;
-  } else if (timeDifferenceInMinutes < 1440) {
-    const hours = Math.floor(timeDifferenceInMinutes / 60);
-    return `${hours}시간 전`;
-  } else {
-    const days = Math.floor(timeDifferenceInMinutes / 1440);
-    return `${days}일 전`;
-  }
-};
+import ChatRoomPreview from './ChatRoomPreview';
 
 const ChatRoomList = ({ userId }: { userId: string }) => {
   const [searchInput, setSearchInput] = useState('');
@@ -84,72 +61,3 @@ const ChatRoomList = ({ userId }: { userId: string }) => {
 };
 
 export default ChatRoomList;
-
-const ChatRoomPreview = ({ room, userId }: { room: RoomWithLastMessageType; userId: string }) => {
-  const { setChatRoom, setChatMessages } = useChatContext();
-  const [searchParams, setSearchParams] = useSearchParams();
-  const currentRoomId = searchParams.get('room_id');
-
-  const handleLeaveRoom = async (room_id: string) => {
-    const confirm = window.confirm('채팅방을 나가시겠습니까?');
-    if (!confirm) return;
-    try {
-      await leaveChatRoom(room_id);
-
-      if (currentRoomId !== room_id) return;
-
-      setChatRoom(null);
-      setChatMessages([]);
-      setSearchParams((prev) => {
-        prev.delete('room_id');
-        return prev;
-      });
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  const usersExcludeMe = room.chat_room_participants.filter((participant) => participant.user_id !== userId);
-  const profile = usersExcludeMe.length > 0 ? usersExcludeMe[0].profiles : undefined;
-
-  return (
-    <S.ChatRoomPreviewContainer key={room.room_id} $isCurrentRoom={room.room_id === currentRoomId}>
-      <S.ChatRoomPreviewLink to={`/chat?room_id=${room.room_id}`}>
-        <S.ProfileImageWrapper>
-          <S.ProfileImage src={profile?.avatar_url || defaultProfileImgUrl} alt="avatar" width={54} height={54} />
-        </S.ProfileImageWrapper>
-        <S.PreviewContent>
-          <S.PreviewTitle>
-            <S.ProfileName>
-              {room.chat_room_participants.find((participant) => participant.user_id !== userId)?.profiles.username || '참가자 없음'}
-              {room.chat_room_participants.length > 2 && ' 외' + room.chat_room_participants.length + '명'}
-            </S.ProfileName>
-            <S.PreviewTime>{room.last_message.length > 0 && <PreviewTime time={room.last_message[0].created_at}></PreviewTime>}</S.PreviewTime>
-          </S.PreviewTitle>
-          <S.PreviewMessage>{room.last_message.length > 0 ? room.last_message[0].content : 'No message'}</S.PreviewMessage>
-        </S.PreviewContent>
-        <div style={{ display: 'none' }}>
-          <button onClick={() => handleLeaveRoom(room.room_id)}>채팅방 나가기</button>
-        </div>
-      </S.ChatRoomPreviewLink>
-    </S.ChatRoomPreviewContainer>
-  );
-};
-
-const PreviewTime = ({ time }: { time: string }) => {
-  const [timeAgo, setTimeAgo] = useState('');
-
-  useEffect(() => {
-    const updateTime = () => {
-      const newTimeAgo = calculateTimeDifference(time);
-      setTimeAgo(newTimeAgo);
-    };
-
-    updateTime();
-    const intervalId = setInterval(updateTime, 60 * 1000);
-
-    return () => clearInterval(intervalId);
-  }, [time]);
-
-  return <>{timeAgo}</>;
-};
